server: refactor sql backend handler
And add some missing tests
