+++ server/src/domain/mod.rs
@@ -3,6 +3,8 @@ pub mod handler;
 pub mod ldap;
 pub mod opaque_handler;
 pub mod sql_backend_handler;
+pub mod sql_group_backend_handler;
 pub mod sql_migrations;
 pub mod sql_opaque_handler;
 pub mod sql_tables;
+pub mod sql_user_backend_handler;
